Fixed regression: fully charged battery shows as depleted.
The regression was introduced by Change Id9fc63f4: Update BatteryMiniIcon code.
This fixes it so the battery shows as full when it's charged entirely and the device
is connected to external power.

Additionally removes a variable which has become useless with the refactory
done in the change Id9fc63f4

Change-Id: I19d634cb174d2dedd1a450e521e4774f07919f51